requirements.txt
Flask==1.1.2
gunicorn==20.1.0
imutils==0.5.3
numpy==1.18.4
opencv-python==4.2.0.32
Pillow==8.1.2
requests==2.25.1
scikit-learn==0.23.1
tensorflow==2.3.1